HW7_Kubeflow_AirFlow_pipelines/PR5_Airflow_Training_Pipeline/requirements.txt
apache-airflow>=2.0.0
torch>=1.7.0
torchvision>=0.8.0
pandas>=1.0.0
numpy>=1.18.0
pillow>=7.0.0
mlflow>=1.15.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
